requirements.txt
openai
python-dotenv
tkinter
